pyproject.toml
[project]
name = "thoughtful-ai-customer-support-agent"
version = "0.1.0"
description = "Thoughtful AI support agent demo (FastAPI + Vite)"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.115.8",
  "uvicorn[standard]==0.34.0",
  "httpx==0.27.2",
  "python-dotenv==1.0.1",
  "scikit-learn==1.5.2",
  "numpy==2.1.3",
]

[tool.uv]
package = false
